refactor: product패키지 위치 변경
