+++ src/main/java/com/hmdandelion/project_1410002/inventory/domian/repository/product/ProductRepository.java
@@ -1,0 +1,12 @@
+package com.hmdandelion.project_1410002.inventory.domian.repository.product;
+
+
+import com.hmdandelion.project_1410002.inventory.domian.entity.product.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.querydsl.QuerydslPredicateExecutor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product,Long>, QuerydslPredicateExecutor<Product> {
+
+}
