+++ src/main/java/com/hmdandelion/project_1410002/inventory/dto/product/request/ProductRequest.java
@@ -1,10 +1,8 @@
-package com.hmdandelion.project_1410002.product.domain.dto.request;
+package com.hmdandelion.project_1410002.inventory.dto.product.request;
 
-import com.hmdandelion.project_1410002.product.domain.type.ProductStatus;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
-import java.time.LocalDateTime;
 @RequiredArgsConstructor
 @Getter
 public class ProductRequest {
