+++ src/main/java/com/hmdandelion/project_1410002/inventory/service/ProductService.java
@@ -1,13 +1,16 @@
-package com.hmdandelion.project_1410002.product.service;
+package com.hmdandelion.project_1410002.inventory.service;
 
 import com.hmdandelion.project_1410002.common.exception.NotFoundException;
 import com.hmdandelion.project_1410002.common.exception.type.ExceptionCode;
-import com.hmdandelion.project_1410002.product.domain.dto.request.ProductRequest;
-import com.hmdandelion.project_1410002.product.domain.dto.response.ProductsResponse;
-import com.hmdandelion.project_1410002.product.domain.entity.Product;
-import com.hmdandelion.project_1410002.product.domain.entity.QProduct;
-import com.hmdandelion.project_1410002.product.domain.repository.ProductRepository;
-import com.hmdandelion.project_1410002.product.domain.type.ProductStatus;
+import com.hmdandelion.project_1410002.inventory.domian.entity.product.Product;
+import com.hmdandelion.project_1410002.inventory.domian.entity.product.QProduct;
+import com.hmdandelion.project_1410002.inventory.domian.repository.product.ProductRepository;
+
+import com.hmdandelion.project_1410002.inventory.domian.type.ProductStatus;
+import com.hmdandelion.project_1410002.inventory.dto.product.request.ProductRequest;
+import com.hmdandelion.project_1410002.inventory.dto.product.response.ProductsResponse;
+
+
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
@@ -71,7 +74,7 @@ public class ProductService {
         product.updateStatus(product);
     }
 
-    public List<Product> searchProducts(Pageable pageable,String productName, String unit,ProductStatus status) {
+    public List<Product> searchProducts(Pageable pageable, String productName, String unit, ProductStatus status) {
 
         QProduct product = QProduct.product;
         BooleanBuilder builder = new BooleanBuilder();
